Refactor role and user repository methods
